[backup] split cli code to modules
Gonna use AsyncReadExt which also have things like read_u32() but async
version, a conflict with std::io::Read
